feat(cosmiconfig): Enable cli to run from anywhere
Bruno's cli is a great tool for quickly dispatching requests from the
terminal. As a bruno cli user, I want to be able to run the command from
any directory, supplying a relative bru file path, and have that request
dispatch as though the command was run within it's containing
collection.

- Adds cosmiconfig to quickly/easily resolve the bruno.json relative to
  the input filename.
- Make some adjustments to related variables to use new config/paths