feat: OpenAPI import: create a default environment
The previous solution constructed a default server baseUrl and directly
concatenated it into each request of the collection. This makes it
difficult to change it later through the UI if it is wrong or if the same
OpenAPI spec can be used in multiple environments.

This solution instead creates a default environment with two variables
"protocol" and "host", which are filled according to the previously
constructed default baseUrl. This makes it very easy to change for all
requests.