Update 'import-collection' logic to handle pre-existing directories
This enables an 'updateExistingCollection' flow which creates any missing
directories while leaving (and traversing) existing directories in place.

For example, if you are working from an OpenAPI file, and the structure
of that file is updated (e.g. by a third party) you can now re-import
that file in order to have any new endpoints be added to an existing
collection. Existing directories, endpoints, or organization are left
untouched.