Fix snippet generation for curl when a binary file is set.
Since files are selected and not enabled, the current code produces an
"Error generating code snippet" error when one is set while exporting
a snippet.

httpsnippet does not directly support --data "@..." when being passed
filepath, so we need to trick it with a formated "text" parameter.

Add the error being sent by httpsnippet to the UI, so that other
issues can be easier debugged.